% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotzeb0912_locs.R
\docType{data}
\name{kotzeb0912_locs}
\alias{kotzeb0912_locs}
\title{Kotzebue 2009-2012 Locations}
\format{Data frame with columns
\describe{
\item{deployid}{unique deployment identifier}
\item{ptt}{argos ptt identifier}
\item{instr}{telemetry instrument type; 'Mk10' or 'SPOT'}
}}
\usage{
kotzeb0912_locs
}
\description{
location data (Argos and GPS) from telemetry devices
}
\examples{
  locations
}
\keyword{datasets}

