% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kotzeb0912_status.R
\docType{data}
\name{kotzeb0912_status}
\alias{kotzeb0912_status}
\title{Kotzebue 2009-2012 Status Messages}
\format{Data frame with columns
\describe{
\item{deployid}{unique deployment identifier}
\item{ptt}{argos ptt identifier}
\item{depthsensor}{depth sensor surface value}
\item{instr}{telemetry instrument type; 'Mk10' or 'SPOT'}
}}
\usage{
kotzeb0912_status
}
\description{
status messages transmitted from telemetry devices
}
\examples{
status
}
\keyword{datasets}

